requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-dotenv==1.1.1
google-genai==1.26.0
PyPDF2==3.0.1
python-docx==1.2.0
Pillow==11.3.0
pytesseract==0.3.13
pdf2image==1.17.0
requests==2.32.4 